+++ src/complex/casinl.c
@@ -1,16 +1,16 @@
 #include "libm.h"
 
 #if LDBL_MANT_DIG == 53 && LDBL_MAX_EXP == 1024
-long double complex casinl(long double complex z)
+double complex casinl(double complex z)
 {
 	return casin(z);
 }
 #else
 // FIXME
-long double complex casinl(long double complex z)
+double complex casinl(double complex z)
 {
-	long double complex w;
-	long double x, y;
+	double complex w;
+	double x, y;
 
 	x = creall(z);
 	y = cimagl(z);
