+++ src/complex/cimagl.c
@@ -1,6 +1,6 @@
 #include "libm.h"
 
-long double (cimagl)(long double complex z)
+double (cimagl)(double complex z)
 {
 	return cimagl(z);
 }
