+++ src/math/coshl.c
@@ -1,17 +1,17 @@
 #include "libm.h"
 
 #if LDBL_MANT_DIG == 53 && LDBL_MAX_EXP == 1024
-long double coshl(long double x)
+double coshl(double x)
 {
 	return cosh(x);
 }
 #elif LDBL_MANT_DIG == 64 && LDBL_MAX_EXP == 16384
-long double coshl(long double x)
+double coshl(double x)
 {
 	union ldshape u = {x};
 	unsigned ex = u.i.se & 0x7fff;
 	uint32_t w;
-	long double t;
+	double t;
 
 	/* |x| */
 	u.i.se = ex;
@@ -40,7 +40,7 @@ long double coshl(long double x)
 }
 #elif LDBL_MANT_DIG == 113 && LDBL_MAX_EXP == 16384
 // TODO: broken implementation to make things compile
-long double coshl(long double x)
+double coshl(double x)
 {
 	return cosh(x);
 }
